sync QA with main

+++ Source/SteamHub.ApiContract/Models/User/CreateUserRequest.cs
@@ -10,9 +10,9 @@ namespace SteamHub.ApiContract.Models.User
 
         public string Password { get; set; }
 
-        public float WalletBalance { get; set; }
+        public decimal WalletBalance { get; set; }
 
-        public float PointsBalance { get; set; }
+        public int PointsBalance { get; set; }
 
         public UserRole UserRole { get; set; }
